# docs/public/gateway_swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ScriptBee Gateway API",
    "description": "API for ScriptBee Gateway service, managing projects, plugins, and analysis.",
    "version": "v2"
  },
  "servers": [
    {
      "url": "http://localhost:5117/"
    }
  ],
  "paths": {
...
    "/api/projects/{projectId}/loaders/{loaderId}/files": {
      "put": {
        "tags": ["Loaders"],
        "summary": "Upload files for a loader",
        "description": "Uploads one or more files to be used by a specific loader in the project.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "loaderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["files"],
                "type": "object",
                "properties": {
                  "files": {
                    "$ref": "#/components/schemas/IFormFileCollection"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadLoaderFilesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/instances/{instanceId}/context": {
      "get": {
        "tags": ["Instances", "Context"],
        "summary": "Get current instance context",
        "description": "Retrieves the current data context for the specified project instance.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectContextResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/instances/{instanceId}/context/generate-classes": {
      "post": {
        "tags": ["Instances", "Context"],
        "summary": "Generate classes for instance context",
        "description": "Generates script classes based on the current data context for the specified project instance and returns them as a stream.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectContextGenerateClassesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/analyses/{analysisId}/results/files/{fileId}": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Download individual analysis result file",
        "description": "Downloads a specific file generated as a result of an analysis.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/analyses/{analysisId}/results/files/download": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Download all analysis result files",
        "description": "Downloads a ZIP archive containing all files generated as a result of an analysis.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/analyses/{analysisId}/scripts/{scriptId}/content": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get analysis script content",
        "description": "Retrieves the actual code content of a script used in a specific analysis.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/analyses/{analysisId}/scripts/{scriptId}": {
      "get": {
        "tags": ["Analysis"],
        "summary": "Get analysis script metadata",
        "description": "Retrieves metadata about a script used in a specific analysis, such as its name and type.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "analysisId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptData"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetProjectContextResponse": {
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectContextSlice"
            }
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "IFormFileCollection": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/IFormFile"
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": ["null", "string"]
          },
          "title": {
            "type": ["null", "string"]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int32"
          },
          "detail": {
            "type": ["null", "string"]
          },
          "instance": {
            "type": ["null", "string"]
          }
        }
      },
...
      "ProjectContextGenerateClassesRequest": {
        "required": ["languages"],
        "type": "object",
        "properties": {
          "languages": {
            "type": ["null", "array"],
            "items": {
              "type": "string"
            }
          },
          "transferFormat": {
            "type": ["null", "string"]
          }
        }
      },
      "ProjectContextSlice": {
        "required": ["model", "pluginIds"],
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "pluginIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ScriptData": {
        "required": ["id", "name", "path", "scriptLanguage", "parameters"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "scriptLanguage": {
            "$ref": "#/components/schemas/ScriptLanguage"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScriptParameter"
            }
          }
        }
      },
      "ScriptLanguage": {
        "required": ["name", "extension"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          }
        }
      },
      "ScriptParameter": {
        "required": ["name", "type", "value"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {}
        }
      },
...
      "UploadLoaderFilesResponse": {
        "required": ["loaderId", "fileNames"],
        "type": "object",
        "properties": {
          "loaderId": {
            "type": "string"
          },
          "fileNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
...
}